segformer-pytorch/requirements.txt
torch
torchvision
tensorboard
numpy==1.17.3
matplotlib==3.1.2
opencv_python==4.1.2.30
tqdm==4.60.0
Pillow==8.2.0
h5py==2.10.0